+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to
 - ✨(backend) add a breadcrumb in the search response
 - ♻️(frontend) move doc action buttons to fix toolbar #2360
 - ♿️(frontend) add aria-hidden to decorative avatar SVGs in share modal #2324
+- 🏗️(frontend) move comments to its own folder feature #2374
 
 ### Fixed
