+++ src/frontend/apps/impress/src/features/docs/doc-comments/api/DocsThreadStoreAuth.tsx
@@ -1,6 +1,6 @@
 import { ThreadStoreAuth } from '@blocknote/core/comments';
 
-import { ClientCommentData, ClientThreadData } from './types';
+import { ClientCommentData, ClientThreadData } from '../types';
 
 export class DocsThreadStoreAuth extends ThreadStoreAuth {
   constructor(
