+++ src/frontend/apps/impress/src/features/docs/doc-comments/index.ts
@@ -1,0 +1,4 @@
+export * from './components/CommentToolbarButton';
+export * from './styles';
+export * from './stores/useCommentSidebarStore';
+export * from './hooks/useComments';
