+++ src/frontend/apps/impress/src/features/docs/doc-comments/stores/useThreadStore.tsx
@@ -1,0 +1,15 @@
+import { create } from 'zustand';
+
+import type { DocsThreadStore } from '../api/DocsThreadStore';
+
+export interface UseThreadStore {
+  threadStore?: DocsThreadStore;
+  setThreadStore: (threadStore: DocsThreadStore | undefined) => void;
+}
+
+export const useThreadStore = create<UseThreadStore>((set) => ({
+  threadStore: undefined,
+  setThreadStore: (threadStore) => {
+    set({ threadStore });
+  },
+}));
