--- src/frontend/apps/impress/src/features/docs/doc-editor/components/comments/index.ts
@@ -1,4 +1,0 @@
-export * from './CommentToolbarButton';
-export * from './styles';
-export * from './useCommentSidebarStore';
-export * from './useComments';
